requirements.txt
click
pyyaml
addict
loguru